Bugfix: Validate field arg type against interface.
This bug occurred when a field argument type was Non-null or List and was
validated to match the type of the interface the object is implementing.

The explanation below just talks about Lists but the same applies for
Non-null too.

The bug is caused because two separate instances of *List were added to
the types list of the root by the parser and later compared during
validation (more on this later). Comparing interfaces checks the
underlying values are identical, which for pointers is false when
non-nil and not pointing to the same struct.

This is fixed by replacing the equality comparison of the two arg types
(using == operator) with an equality function that checks the types
match all the way down to the base type name. This equality function
will work even if the two operands' underlying types are pointers to
separate structs.

I also considered making the parser only add one instance of the type to
the types list and look it up later. This would allow == comparison of
the underlying pointers however this would cause problems with reporting
line/column number in error messages. I noticed that the types added to
the type list were no longer in the list after the validation step of
the SDL parser, so I opted to go for the type equality function.