Bugfix: Race conditions during reflection resolution.
The previous bugfix did not address concurrent writes happening on
Field Definitions and Objects when resolving using the reflection
approach.

These were fixed by the addition of mutexes to synchronize the
reads/writes. From what I can tell the writes to the fields during
resolution cannot efficiently be done ahead of time because the
reflection objects may be created during the process of resolving (e.g.
after reading from a database). If there is a way to map GraphQL types
to a sample of their reflection object type ahead of time, the writes
(and need for synchronization) during resolution could be removed.