Add Extensions to Errors
The GraphQL Specification allows for an "extensions" field to be added
to errors for use by implementers to extend the error information
available to the user. This is useful for providing error codes or other
structured data.

This change adds support for error extensions. Implementers who return a
*ggql.Error from a resolver method containing Extensions data will now
see this data returned to the API caller.

See spec: http://spec.graphql.org/June2018/#sec-Errors